feature(datavalues): add series filter
