+++ common/datavalues/src/columns/boolean/mod.rs
@@ -15,8 +15,6 @@
 use std::sync::Arc;
 
 use common_arrow::arrow::array::*;
-use common_arrow::arrow::bitmap::utils::BitChunkIterExact;
-use common_arrow::arrow::bitmap::utils::BitChunksExact;
 use common_arrow::arrow::bitmap::Bitmap;
 use common_arrow::arrow::bitmap::MutableBitmap;
 use common_arrow::arrow::datatypes::DataType as ArrowType;
@@ -109,42 +107,7 @@ impl Column for BooleanColumn {
     }
 
     fn filter(&self, filter: &BooleanColumn) -> ColumnRef {
-        let selected = filter.values().len() - filter.values().null_count();
-        if selected == self.len() {
-            return Arc::new(self.clone());
-        }
-        let mut bitmap = MutableBitmap::with_capacity(selected);
-        let (value_slice, _, value_length) = self.values().as_slice();
-        let (slice, _, length) = filter.values().as_slice();
-
-        let mut chunks = BitChunksExact::<u64>::new(value_slice, value_length);
-        let mut mask_chunks = BitChunksExact::<u64>::new(slice, length);
-
-        chunks
-            .by_ref()
-            .zip(mask_chunks.by_ref())
-            .for_each(|(chunk, mut mask)| {
-                while mask != 0 {
-                    let n = mask.trailing_zeros() as usize;
-                    let value: bool = chunk & (1 << n) != 0;
-                    bitmap.push(value);
-                    mask = mask & (mask - 1);
-                }
-            });
-
-        chunks
-            .remainder_iter()
-            .zip(mask_chunks.remainder_iter())
-            .for_each(|(value, is_selected)| {
-                if is_selected {
-                    bitmap.push(value);
-                }
-            });
-
-        let col = BooleanColumn {
-            values: bitmap.into(),
-        };
-        Arc::new(col)
+        Series::filter_column(self, filter)
     }
 
     fn scatter(&self, indices: &[usize], scattered_size: usize) -> Vec<ColumnRef> {
@@ -203,6 +166,11 @@ impl ScalarColumn for BooleanColumn {
     type OwnedItem = bool;
     type RefItem<'a> = bool;
     type Iterator<'a> = BitmapValuesIter<'a>;
+
+    #[inline]
+    fn clone_column(&self) -> ColumnRef {
+        Arc::new(self.clone())
+    }
 
     #[inline]
     fn get_data(&self, idx: usize) -> Self::RefItem<'_> {
