+++ common/datavalues/src/columns/mod.rs
@@ -21,6 +21,7 @@ mod column;
 mod column_with_field;
 mod const_;
 mod eq;
+mod filter;
 #[allow(clippy::ptr_arg)]
 mod group_hash;
 mod mutable;
@@ -39,6 +40,7 @@ pub use builder::*;
 pub use column::*;
 pub use column_with_field::*;
 pub use const_::*;
+pub use filter::*;
 pub use group_hash::GroupHash;
 pub use mutable::*;
 pub use null::*;
