+++ common/datavalues/src/columns/object/mod.rs
@@ -226,6 +226,11 @@ where T: Scalar<ColumnType = Self> + ObjectType
     type Iterator<'a> = ObjectValueIter<'a, T>;
 
     #[inline]
+    fn clone_column(&self) -> ColumnRef {
+        Arc::new(self.clone())
+    }
+
+    #[inline]
     fn get_data(&self, idx: usize) -> Self::RefItem<'_> {
         self.values[idx].as_scalar_ref()
     }
